+++ CHANGES.rst
@@ -12,6 +12,11 @@ esa/xmm-newton
 Service fixes and enhancements
 ------------------------------
 
+casda
+^^^^^
+
+- Add ability to stage and download ASKAP data. [#1706]
+
 gaia
 ^^^^